Fastapi/requirements.txt
# Core
fastapi>=0.110.0,<0.112.0
uvicorn[standard]>=0.25.0,<0.26.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0

# Database & Vector
pymysql>=1.1.0,<2.0.0
sqlalchemy>=2.0.0,<3.0.0
tidb-vector>=0.1.0,<0.2.0

# AI & Embeddings
sentence-transformers>=2.2.2,<3.0.0
torch>=2.0.0,<3.0.0
transformers>=4.30.0,<5.0.0

# Utilities
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
requests>=2.28.0,<3.0.0
httpx>=0.24.0,<0.28.0
python-multipart>=0.0.5,<0.1.0
python-dateutil>=2.8.2,<3.0.0
pillow>=10.0.0,<11.0.0
python-magic>=0.4.27,<0.5.0
tqdm>=4.65.0,<5.0.0
numpy>=1.24.0,<2.0.0

# Optional
aiofiles>=23.1.0,<24.0.0  # For async file operations
emails>=0.6.0,<1.0.0      # Lightweight SMTP helper